Mostly working test cases
