+++ language/models/cache.ts
@@ -5,7 +5,7 @@ const newInds = () => {
   return [...Array(98).keys(), `LR`, `KL`].map(val => `IN${val.toString().padStart(2, `0`)}`).map(ind => {
     const indDef = new Declaration(`variable`);
     indDef.name = ind;
-    indDef.keywords = [`IND`];
+    indDef.keyword = {IND: true};
     return indDef;
   })
 };
@@ -148,21 +148,6 @@ export default class Cache {
     } else {
       return null;
     }
-  }
-
-  clearReferences() {
-    const fileStructs = this.files.map(file => file.subItems).flat();
-
-    [...fileStructs, ...this.parameters, ...this.constants, ...this.files, ...this.procedures, ...this.subroutines, ...this.variables, ...this.structs].forEach(def => {
-      def.references = [];
-      def.subItems.forEach(sub => sub.references = []);
-    });
-
-    this.procedures.forEach(proc => {
-      if (proc.scope) {
-        proc.scope.clearReferences();
-      }
-    });
   }
 
   findDefinition(lineNumber, word) {
